Refactor caches and Supabase helpers
